requirements.txt
streamlit
requests
numpy
newspaper3k
scikit-learn
langchain
mailgun
tiktoken
openai
lxml